requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
db-dtypes==1.1.1
decorator==5.1.1
elevenlabs==0.2.26
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.104.1
frozenlist==1.4.0
google-api-core==2.14.0
google-auth==2.24.0
google-auth-oauthlib==1.1.0
google-cloud-bigquery==3.13.0
google-cloud-bigquery-storage==2.23.0
google-cloud-core==2.3.3
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.3
grpcio-status==1.59.3
h11==0.14.0
idna==3.6
ipython==8.12.3
jedi==0.19.1
matplotlib-inline==0.1.6
multidict==6.0.4
numpy==1.24.4
oauthlib==3.2.2
openai==0.28.1
packaging==23.2
pandas==2.0.3
pandas-gbq==0.19.2
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
prompt-toolkit==3.0.41
proto-plus==1.22.3
protobuf==4.25.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.2
pydantic-core==2.14.5
pydata-google-auth==1.8.2
pygments==2.17.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytube==15.0.0
pytz==2023.3.post1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scrapetube==2.5.1
six==1.16.0
sniffio==1.3.0
stack-data==0.6.3
starlette==0.27.0
tiktoken==0.5.2
tqdm==4.66.1
traitlets==5.14.0
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
uvicorn==0.24.0.post1
wcwidth==0.2.12
websockets==12.0
yarl==1.9.3